+++ main.py
@@ -62,7 +62,7 @@ import re
 
 
 # 注册插件，指定插件名、作者、描述和版本号
-@register("mrfzccl", "Lishining", "你知道的,我一直是明日方舟高手", "1.2.42")
+@register("mrfzccl", "Lishining", "你知道的,我一直是明日方舟高手", "1.2.44")
 class Mrfzccl(Star):
     _question_candidate_names: np.ndarray
     _question_candidate_urls: List[List[str]]
@@ -272,7 +272,6 @@ class Mrfzccl(Star):
         self.renderer = renderer_cls(
             output_dir=str(self.img_tmp_path),
             t2i_enabled=t2i_config.enabled,
-            t2i_endpoint=t2i_config.endpoint,
             t2i_max_concurrent=t2i_config.max_concurrent,
             html_render_func=self.html_render,
         )